Implement a lot of the breakpoint (and watchpoint) logic for the JTAG ICE mkII.  This is the code base I've been testing for a while, and it works at least as well as the mkI implementation.
I completely dropped the code related to the "step-over-range-packet"
extension, supposedly the "e" packet, as it appears the respective GDB
code never made it into the official GDB sources.  As it had not been
used for quite some time, dropping it now untangles some parts of the
code.

The device descriptor table now gets an additional field with
per-device flags.  The first purpose was to mark the ATmega128 as `no
software breakpoints available' as it is known to not implement the
BREAK instruction.  Further bits can be allocated as needed.

jtag2rw.cc now contains a hack that tries to identify a "load"
command, and chip-erases the device.  This doesn't work yet as
intented, but might be a start.  It appears to me the "load" command
never worked before (as it didn't handle any form of erases), but
then, perhaps the mkI implementation of the ICE is just different
here.


git-svn-id: https://svn.code.sf.net/p/avarice/code/trunk@113 a7e36acc-7495-4a18-91ca-7b9a7881be12